Add router and update APP CSS to support the new route
