+++ src/components/Home/Home.tsx
@@ -1,0 +1,5 @@
+const Home = () => {
+  return <></>;
+};
+
+export default Home;
